scripts/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
geojson==2.5.0
idna==3.3
psycopg2-binary==2.9.1
requests==2.26.0
urllib3==1.26.7
